Update team_list_dict
I put each team along with the number of wins they had into their own separate dictionary. Then I put those dictionaries into a list so that when prompted for a team/country the program will iterate through the list of dictionaries and spit out the number of world cup wins the team/country has. I will add more items for the engine to spit out for each country such as year(s) they won, the team they played, the score, etc. I also want to do something like this for football (as I am in a fantasy football league) to help me get a quick look at the teams with the best defense (against the run, in the air) and offense (on the ground, in the air) so that I can figure out who to sit/start and who to potentially trade for (the trading part would work better if I could create an algorithm to look at future matchups for a specific team and determine difficulty of their future schedule, but I will tackle that once I get a better feel for this stuff – one thing at a time)